requirements.txt
faster-whisper>=1.0.0
sounddevice>=0.4.6
numpy>=1.24
keyboard>=0.13.5
pyperclip>=1.8.2
winotify>=1.1.0
pillow>=10.0.0  # build-time only, for make_icon.py
